feat: :sparkles: Basic schema analysis
