+++ server/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from routers import data, migration, updater
+from routers import data, migration, updater, schema
 
 app = FastAPI(
     title="Tweaked.ai",
@@ -10,6 +10,7 @@ app = FastAPI(
 app.include_router(data.router, tags=["data"])
 app.include_router(migration.router, tags=["migrations"])
 app.include_router(updater.router, tags=["updater"])
+app.include_router(schema.router, tags=["schema"])
 
 origins = ["*"]
